working on faculty edit

+++ src/main/java/DAO/mapper/ObjectMapper.java
@@ -8,6 +8,6 @@ public interface ObjectMapper<T> {
 
     T extractFromResultSet(ResultSet rs) throws SQLException;
 
-    T makeUnique(Map<Integer, T> cache,
+    T makeUnique(Map<Long, T> cache,
                  T entity);
 }
